% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPDM.R
\name{invTrans}
\alias{invTrans}
\title{Transformation function}
\usage{
invTrans(v)
}
\arguments{
\item{v}{Real, positive vector}
}
\value{
Real vector, containing the result of inverse transformation function.
}
\description{
Applies the inverse transformation \eqn{t^{-1}(v) = \log(\exp{(v)} -1)}
}
\details{
Transformation from real, positive vector in real vector under preservation of frechet-distribution.
}
\references{
Cooley & Thibaud (2019) <doi:10.1093/biomet/asz028>
}
\seealso{
\link{svd.tpdm}, \link{pca.tpdm}
}
\author{
Yuing Jiang, Dan Cooley
}
