% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precipGER}
\alias{precipGER}
\title{daily Precipitation over Southern Germany}
\format{
A list containing containing
\itemize{
    \item pr: data-array
    \item date: time-information
    \item lon,lat: longitude & latitude information
}
}
\source{
Quelle: Deutscher Wetterdienst
}
\usage{
data(precipGER)
}
\description{
Daily Precipitation at several stations in Germany
}
\details{
Daily Precipitation Data

Daily precipitation data from several wather station in southern Germany (longitude <50) over the years 2000-2019.
The data has been downloaded from opendata server of german weather service (https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/historical/).
}
\keyword{datasets}
