% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPDM.R
\name{to.alpha.2}
\alias{to.alpha.2}
\title{Probability integral transformation}
\usage{
to.alpha.2(data, orig = NULL)
}
\arguments{
\item{data}{A t x n dimensional, numeric Data-matrix with t: Number of time steps and n: Number of grid points/stations}

\item{orig}{If known: original distribution of data (currently implemented: 'normal' or 'gamma'), else: NULL}
}
\value{
Data-matrix of same dimension as 'data', but in Frechet-margins with tail-index 2
}
\description{
Performs transformation to make all of the margins follow a Frechet distribution with tail-index alpha = 2.
}
