% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPDM.R
\name{trans}
\alias{trans}
\title{transformation function}
\usage{
trans(x)
}
\arguments{
\item{x}{Real vector}
}
\value{
Real, positive vector, containing the result of transformation function.
}
\description{
Applies the transformation \eqn{t(x) = \log(1+\exp{(x)})}
}
\details{
Transformation from real vector in real, positive vector under preservation of Frechet-distribution.
}
\references{
Cooley & Thibaud (2019) <doi:10.1093/biomet/asz028>
}
\seealso{
\link{svd.tpdm}, \link{pca.tpdm}
}
\author{
Yuing Jiang, Dan Cooley
}
